refactor(credential): rename 1.4.2 folders to 1.4.4
