+++ src/credential/issuance/v1.4.4/04-authorize-access.ts
