+++ src/credential/issuance/v1.4.4/06-verify-and-parse-credential.ts
