+++ src/credential/offer/v1.4.4/02-extract-grant-details.ts
