+++ src/credential/offer/v1.4.4/03-validate-credential-offer.ts
