+++ src/credential/offer/v1.4.4/__tests__/01-resolve-credential-offer.test.ts
