+++ src/credential/offer/v1.4.4/__tests__/02-extract-grant-details.test.ts
