+++ src/credential/offer/v1.4.4/__tests__/03-validate-credential-offer.test.ts
