+++ src/credential/presentation/index.ts
@@ -2,4 +2,4 @@ import * as Errors from "./common/errors";
 export { Errors };
 export type * from "./api";
 export { RemotePresentation as V1_0_0 } from "./v1.0.0";
-export { RemotePresentation as V1_4_2 } from "./v1.4.2";
+export { RemotePresentation as V1_4_4 } from "./v1.4.4";
