+++ src/credential/presentation/v1.4.4/05-verify-request-object.ts
