+++ src/credential/presentation/v1.4.4/07-send-authorization-response.ts
