+++ src/credential/trustmark/v1.4.4/get-credential-trustmark.ts
