+++ src/credentials-catalogue/index.ts
@@ -1,3 +1,3 @@
 export type * from "./api";
 export { CredentialsCatalogue as V1_0_0 } from "./v1.0.0";
-export { CredentialsCatalogue as V1_4_2 } from "./v1.4.2";
+export { CredentialsCatalogue as V1_4_4 } from "./v1.4.2";
